requirements.txt
python-keystoneclient==2.0.0
python-heatclient==0.8.0
six==1.10.0
celery==3.1.18
paramiko==1.9.0
-e git+https://github.com/edx/xblock-utils.git@v1.0.2#egg=xblock-utils==v1.0.2
-e .
